Box on a Ramp Sketch
This is the beginning of a new sim that is still heavily under development. The idea is to show a FBD of a box on a ramp with a slider to adjust the angle of the ramp and show the corresponding changes in components of the FBD. Currently there is an adjustable ramp and a box that follows accordingly. FBD to come soon