Create sketch.js
This is an updated version of a box on a ramp. The FBD is still not developed for this sim yet; currently exploring using either the Arrow() function of fbd() function from science.js. The ramp now has a slider to adjust the angle and output the corresponding value to user. There is also now a box on the ramp that obeys gravity and friction (to an extent...) when the angle of ramp is changed. The box will start to slide down the ramp at higher angles, but it will slide uninhibited by kinetic friction (which will change soon...).